feat(ui): replace inline search functionality with dedicated SearchBar component
